feat(software-demo): move plumbing implementation from testing-only module to part of the implementation
This makes sense since the crate has been renamed to software-demo, and
it also illustrates better how extenders can compose also with
functionality they provide themselves.
