+++ embedded-cal-software-demo/src/hash_plumbing.rs
@@ -2,30 +2,149 @@
 // SPDX-FileCopyrightText: Inria-AIO, Cryspen, and Christian Amsüss
 //! Implementations of SHA2 plumbing
 //!
-//! Currently, this is only passing through.
+//! Whether this passes everything through or implements everything depends on
+//! `EC::IMPLEMENT_SHA2PLUMBING`.
 
 use super::{Extender, ExtenderConfig};
 use embedded_cal::plumbing::hash::Sha2Short;
 
-impl<EC: ExtenderConfig> Sha2Short for Extender<EC> {
-    const SUPPORTED: bool = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SUPPORTED;
-    const SEND_PADDING: bool = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SEND_PADDING;
-    const FIRST_CHUNK_SIZE: usize =
-        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::FIRST_CHUNK_SIZE;
-    const UPDATE_MULTICHUNK: bool =
-        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::UPDATE_MULTICHUNK;
+pub enum State<EC: ExtenderConfig> {
+    // FIXME: As with the Hash types, this should effectively just be transparent around whichever
+    // is the one actually used version.
+    SoftwareSha2Short([u32; 8]),
+    Direct(<EC::Base as Sha2Short>::State),
+}
 
-    type State = <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::State;
+impl<EC: ExtenderConfig> Clone for State<EC> {
+    // This is the default implemnentation, but we can't derive it because EC is not clone. (We
+    // don't expect it to, but we'd need "minimal derives" in Rust to make it derivable).
+    fn clone(&self) -> Self {
+        match self {
+            Self::SoftwareSha2Short(arg) => Self::SoftwareSha2Short(*arg),
+            Self::Direct(arg) => Self::Direct(arg.clone()),
+        }
+    }
+}
+
+impl<EC: ExtenderConfig> Sha2Short for Extender<EC> {
+    const SUPPORTED: bool = if EC::IMPLEMENT_SHA2SHORT_PLUMBING {
+        true
+    } else {
+        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SUPPORTED
+    };
+    const SEND_PADDING: bool = if EC::IMPLEMENT_SHA2SHORT_PLUMBING {
+        true
+    } else {
+        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::SEND_PADDING
+    };
+    const FIRST_CHUNK_SIZE: usize = if EC::IMPLEMENT_SHA2SHORT_PLUMBING {
+        64
+    } else {
+        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::FIRST_CHUNK_SIZE
+    };
+    const UPDATE_MULTICHUNK: bool = if EC::IMPLEMENT_SHA2SHORT_PLUMBING {
+        false
+    } else {
+        <EC::Base as embedded_cal::plumbing::hash::Sha2Short>::UPDATE_MULTICHUNK
+    };
+
+    type State = State<EC>;
 
     fn init(&mut self, variant: embedded_cal::plumbing::hash::Sha2ShortVariant) -> Self::State {
-        self.0.init(variant)
+        if !EC::IMPLEMENT_SHA2SHORT_PLUMBING {
+            return State::Direct(self.0.init(variant));
+        }
+
+        // FIXME
+        if !matches!(
+            variant,
+            embedded_cal::plumbing::hash::Sha2ShortVariant::Sha256
+        ) {
+            todo!();
+        }
+
+        State::SoftwareSha2Short([
+            0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a, 0x510e527f, 0x9b05688c, 0x1f83d9ab,
+            0x5be0cd19,
+        ])
     }
 
+    #[allow(non_snake_case, reason = "folling algorithm convention")]
     fn update(&mut self, instance: &mut Self::State, data: &[u8]) {
-        self.0.update(instance, data)
+        let instance = match instance {
+            State::Direct(instance) => return self.0.update(instance, data),
+            State::SoftwareSha2Short(instance) => instance,
+        };
+
+        assert!(data.len() == 64, "Not feeding exactly 512bit");
+
+        let mut w: [u32; 64] = [0; _]; // or uninit
+        for (i, chunkword) in data.as_chunks::<4>().0.iter().enumerate() {
+            w[i] = u32::from_be_bytes(*chunkword);
+        }
+
+        for i in 16..64 {
+            let s0 = w[i - 15].rotate_right(7) ^ w[i - 15].rotate_right(18) ^ (w[i - 15] >> 3);
+            let s1 = w[i - 2].rotate_right(17) ^ w[i - 2].rotate_right(19) ^ (w[i - 2] >> 10);
+            w[i] = w[i - 16]
+                .wrapping_add(s0)
+                .wrapping_add(w[i - 7])
+                .wrapping_add(s1);
+        }
+
+        let [mut a, mut b, mut c, mut d, mut e, mut f, mut g, mut h] = *instance;
+
+        for i in 0..64 {
+            let S1 = e.rotate_right(6) ^ e.rotate_right(11) ^ e.rotate_right(25);
+            let ch = (e & f) ^ (!e & g);
+            let temp1 = h
+                .wrapping_add(S1)
+                .wrapping_add(ch)
+                .wrapping_add(k[i])
+                .wrapping_add(w[i]);
+            let S0 = a.rotate_right(2) ^ a.rotate_right(13) ^ a.rotate_right(22);
+            let maj = (a & b) ^ (a & c) ^ (b & c);
+            let temp2 = S0.wrapping_add(maj);
+
+            h = g;
+            g = f;
+            f = e;
+            e = d.wrapping_add(temp1);
+            d = c;
+            c = b;
+            b = a;
+            a = temp1.wrapping_add(temp2);
+        }
+
+        for (big_h, add_this) in instance.iter_mut().zip([a, b, c, d, e, f, g, h]) {
+            *big_h = big_h.wrapping_add(add_this);
+        }
     }
 
     fn finalize(&mut self, instance: Self::State, last_chunk: &[u8], target: &mut [u8]) {
-        self.0.finalize(instance, last_chunk, target)
+        let instance = match instance {
+            State::Direct(instance) => return self.0.finalize(instance, last_chunk, target),
+            State::SoftwareSha2Short(instance) => instance,
+        };
+        assert!(
+            last_chunk.is_empty(),
+            "Implementation requested that padding be sent and didn't set a fancy first block size, so we're not expecting a non-empty last chunk"
+        );
+
+        for (i, word) in instance.into_iter().enumerate() {
+            target[4 * i..][..4].copy_from_slice(&word.to_be_bytes());
+        }
     }
 }
+
+#[allow(non_upper_case_globals, reason = "folling algorithm convention")]
+const k: [u32; 64] = [
+    0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
+    0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3, 0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174,
+    0xe49b69c1, 0xefbe4786, 0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
+    0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147, 0x06ca6351, 0x14292967,
+    0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13, 0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85,
+    0xa2bfe8a1, 0xa81a664b, 0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
+    0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a, 0x5b9cca4f, 0x682e6ff3,
+    0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208, 0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2,
+];
