+++ embedded-cal-software-demo/src/hkdf.rs
@@ -5,19 +5,19 @@
 
 #[cfg(test)]
 mod tests {
-    use crate::tests::dummy_sha256;
     use crate::{Extender, ExtenderConfig};
 
     struct ImplementSha256Short;
 
     impl ExtenderConfig for ImplementSha256Short {
         const IMPLEMENT_SHA2SHORT: bool = true;
-        type Base = dummy_sha256::DummySha256;
+        const IMPLEMENT_SHA2SHORT_PLUMBING: bool = true;
+        type Base = embedded_cal::empty::EmptyCal;
     }
 
     #[test]
-    fn test_hkdf_sha256_on_dummy() {
-        let mut cal = Extender::<ImplementSha256Short>(dummy_sha256::DummySha256::new());
+    fn test_hkdf_sha256() {
+        let mut cal = Extender::<ImplementSha256Short>(embedded_cal::empty::EmptyCal);
         testvectors::test_hkdf_sha256(&mut cal);
     }
 }
