--- embedded-cal-software-demo/src/tests/dummy_sha256.rs
@@ -1,161 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-// SPDX-FileCopyrightText: Inria-AIO, Cryspen, and Christian Amsüss
-//! Testing-only module with a manual version of SHA256 that does *not* do the padding (in order to
-//! test the `fullblock_nonfinishing_from_cose_number()` algorithms)
-#![allow(
-    non_snake_case,
-    non_upper_case_globals,
-    reason = "folling algorithm convention"
-)]
-
-use embedded_cal::empty::EmptyCal;
-
-/// A minimal testable version of SHA256-but-no-blocks-or-dummy.
-///
-/// All implementation follows the Wikipedia pseudocode.
-pub struct DummySha256(EmptyCal);
-
-impl DummySha256 {
-    pub fn new() -> Self {
-        Self(EmptyCal)
-    }
-}
-
-const k: [u32; 64] = [
-    0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
-    0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3, 0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174,
-    0xe49b69c1, 0xefbe4786, 0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
-    0x983e5152, 0xa831c66d, 0xb00327c8, 0xbf597fc7, 0xc6e00bf3, 0xd5a79147, 0x06ca6351, 0x14292967,
-    0x27b70a85, 0x2e1b2138, 0x4d2c6dfc, 0x53380d13, 0x650a7354, 0x766a0abb, 0x81c2c92e, 0x92722c85,
-    0xa2bfe8a1, 0xa81a664b, 0xc24b8b70, 0xc76c51a3, 0xd192e819, 0xd6990624, 0xf40e3585, 0x106aa070,
-    0x19a4c116, 0x1e376c08, 0x2748774c, 0x34b0bcb5, 0x391c0cb3, 0x4ed8aa4a, 0x5b9cca4f, 0x682e6ff3,
-    0x748f82ee, 0x78a5636f, 0x84c87814, 0x8cc70208, 0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2,
-];
-
-impl embedded_cal::Cal for DummySha256 {
-    type DhProvider = EmptyCal;
-    type AeadProvider = EmptyCal;
-    type HashProvider = EmptyCal;
-    type HmacProvider = EmptyCal;
-
-    fn dh(&mut self) -> &mut Self::DhProvider {
-        &mut self.0
-    }
-
-    fn aead(&mut self) -> &mut Self::AeadProvider {
-        &mut self.0
-    }
-
-    fn hash(&mut self) -> &mut Self::HashProvider {
-        &mut self.0
-    }
-
-    fn hmac(&mut self) -> &mut Self::HmacProvider {
-        &mut self.0
-    }
-}
-
-impl embedded_cal::plumbing::Plumbing for DummySha256 {}
-
-impl embedded_cal::plumbing::hash::Hash for DummySha256 {}
-
-impl embedded_cal::plumbing::hash::Sha2Short for DummySha256 {
-    const SUPPORTED: bool = true;
-    const SEND_PADDING: bool = true;
-    const FIRST_CHUNK_SIZE: usize = 64;
-    const UPDATE_MULTICHUNK: bool = false;
-
-    type State = [u32; 8];
-
-    fn init(&mut self, variant: embedded_cal::plumbing::hash::Sha2ShortVariant) -> Self::State {
-        // FIXME
-        if !matches!(
-            variant,
-            embedded_cal::plumbing::hash::Sha2ShortVariant::Sha256
-        ) {
-            todo!();
-        }
-
-        [
-            0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a, 0x510e527f, 0x9b05688c, 0x1f83d9ab,
-            0x5be0cd19,
-        ]
-    }
-
-    fn update(&mut self, instance: &mut Self::State, data: &[u8]) {
-        assert!(data.len() == 64, "Not feeding exactly 512bit");
-
-        let mut w: [u32; 64] = [0; _]; // or uninit
-        for (i, chunkword) in data.as_chunks::<4>().0.iter().enumerate() {
-            w[i] = u32::from_be_bytes(*chunkword);
-        }
-
-        for i in 16..64 {
-            let s0 = w[i - 15].rotate_right(7) ^ w[i - 15].rotate_right(18) ^ (w[i - 15] >> 3);
-            let s1 = w[i - 2].rotate_right(17) ^ w[i - 2].rotate_right(19) ^ (w[i - 2] >> 10);
-            w[i] = w[i - 16]
-                .wrapping_add(s0)
-                .wrapping_add(w[i - 7])
-                .wrapping_add(s1);
-        }
-
-        let [mut a, mut b, mut c, mut d, mut e, mut f, mut g, mut h] = *instance;
-
-        for i in 0..64 {
-            let S1 = e.rotate_right(6) ^ e.rotate_right(11) ^ e.rotate_right(25);
-            let ch = (e & f) ^ (!e & g);
-            let temp1 = h
-                .wrapping_add(S1)
-                .wrapping_add(ch)
-                .wrapping_add(k[i])
-                .wrapping_add(w[i]);
-            let S0 = a.rotate_right(2) ^ a.rotate_right(13) ^ a.rotate_right(22);
-            let maj = (a & b) ^ (a & c) ^ (b & c);
-            let temp2 = S0.wrapping_add(maj);
-
-            h = g;
-            g = f;
-            f = e;
-            e = d.wrapping_add(temp1);
-            d = c;
-            c = b;
-            b = a;
-            a = temp1.wrapping_add(temp2);
-        }
-
-        for (big_h, add_this) in instance.iter_mut().zip([a, b, c, d, e, f, g, h]) {
-            *big_h = big_h.wrapping_add(add_this);
-        }
-    }
-
-    fn finalize(&mut self, instance: Self::State, last_chunk: &[u8], target: &mut [u8]) {
-        assert!(
-            last_chunk.is_empty(),
-            "Implementation requested that padding be sent and didn't set a fancy first block size, so we're not expecting a non-empty last chunk"
-        );
-
-        for (i, word) in instance.into_iter().enumerate() {
-            target[4 * i..][..4].copy_from_slice(&word.to_be_bytes());
-        }
-    }
-}
-
-impl embedded_cal::plumbing::ec::Ec for DummySha256 {
-    const MAX_SCALAR_LENGTH: usize = 0;
-
-    type PrimitivesP256 = EmptyCal;
-    type PrimitivesX25519 = EmptyCal;
-    type PrimitivesX448 = EmptyCal;
-
-    fn p256(&mut self) -> &mut Self::PrimitivesP256 {
-        &mut self.0
-    }
-
-    fn x25519(&mut self) -> &mut Self::PrimitivesX25519 {
-        &mut self.0
-    }
-
-    fn x448(&mut self) -> &mut Self::PrimitivesX448 {
-        &mut self.0
-    }
-}
