+++ embedded-cal-stm32wba55/tests/integration.rs
@@ -9,6 +9,7 @@ use panic_probe as _;
 struct ImplementSha256Short;
 impl embedded_cal_software_demo::ExtenderConfig for ImplementSha256Short {
     const IMPLEMENT_SHA2SHORT: bool = true;
+    const IMPLEMENT_SHA2SHORT_PLUMBING: bool = false;
     type Base = embedded_cal_stm32wba55::Stm32wba55Cal;
 }
